ref: remove P2PEnabledConference
